lorenzo-app/langchain-ask-pdf/requirements.txt
constants==0.6.0
langchain==0.0.308
pandas==2.0.3
python-dotenv==1.0.0
streamlit==1.27.2
